+++ .github/workflows/node.yml
@@ -19,3 +19,4 @@ jobs:
       - run: npm install
       - run: npm run lint
       - run: npm run build
+      - run: npm run test
