+++ package.json
@@ -4,7 +4,7 @@
   "description": "A GraphQL wrapper for the tempus.xyz api",
   "main": "dist/schema.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "npm run build && jest",
     "graphiql": "npm run build && node ./scripts/graphiql.js",
     "build": "babel lib -d dist",
     "prepare": "npm run build",
@@ -22,6 +22,11 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/arispoloway/tempus-api-graphql.git"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "lib"
+    ]
   },
   "dependencies": {
     "graphql": "^15.3.0",
@@ -41,6 +46,7 @@
     "eslint-plugin-prettier": "^3.1.4",
     "express": "^4.17.1",
     "express-graphql": "^0.11.0",
+    "jest": "^26.2.2",
     "prettier": "^2.0.5"
   }
 }
